test(sdk-trace-base): move util.inspect tests to test/node
The karma browser harness loads tests from test/common, and the browser
shim of `util.inspect` does not honor `Symbol.for('nodejs.util.inspect.custom')`
the way Node does — assertions on the class-tagged prefix were failing.
Move the inspect tests under test/node so only the Node test runner picks
them up; the inspect symbol itself is a no-op in browsers, which is the
intended behaviour.
